Visualize the angles between joints across frames

Visualize the angles between joints across frames

Usage

plot_degrees(
  data,
  x,
  y,
  joint_ids,
  frames,
  inv_joint_angles = NULL,
  smooth = NULL,
  ...,
  plot = TRUE
)

Arguments

data

Dataset with motion capture data.

x

Unquoted variable name of the x coordinate variable.

y

Unquoted variable name of the y coordinate variable.

joint_ids

Unquoted variable name with the joint ids.

frames

Unquoted variable name with the frame ids for each joint.

inv_joint_angles

Character. Levels of the joint_ids to invert the angles (i.e., negative to positive angles and viceversa).

smooth

Character. Method to use to draw a smooth line interpolating the data points. Options are "loess" or "gam". (defaults to NULL).

...

Further arguments passed to ggplot2::geom_smooth().

plot

Logical. Wheather to obtain the plot (TRUE) or the data with angle measurements (F).

Value

ggplot2 object